+++ crates/rpc/src/cairo/ext_py.rs
@@ -521,7 +521,7 @@ mod tests {
 
         assert_eq!(
             at_block_fee,
-            crate::rpc::v01::types::reply::FeeEstimate {
+            crate::v01::types::reply::FeeEstimate {
                 consumed: H256::from_low_u64_be(0x55a),
                 gas_price: H256::from_low_u64_be(1),
                 fee: H256::from_low_u64_be(0x55a),
@@ -544,7 +544,7 @@ mod tests {
 
         assert_eq!(
             current_fee,
-            crate::rpc::v01::types::reply::FeeEstimate {
+            crate::v01::types::reply::FeeEstimate {
                 consumed: H256::from_low_u64_be(0x55a),
                 gas_price: H256::from_low_u64_be(10),
                 fee: H256::from_low_u64_be(0x3584),
@@ -612,7 +612,7 @@ mod tests {
 
         assert_eq!(
             at_block_fee,
-            crate::rpc::v01::types::reply::FeeEstimate {
+            crate::v01::types::reply::FeeEstimate {
                 consumed: H256::from_low_u64_be(0xa2c),
                 gas_price: H256::from_low_u64_be(1),
                 fee: H256::from_low_u64_be(0xa2c),
