+++ crates/rpc/src/v02/method/add_declare_transaction.rs
@@ -1,11 +1,11 @@
-use crate::rpc::v02::types::request::BroadcastedDeclareTransaction;
-use crate::rpc::v02::RpcContext;
+use crate::v02::types::request::BroadcastedDeclareTransaction;
+use crate::v02::RpcContext;
 use pathfinder_common::{ClassHash, StarknetTransactionHash};
 use starknet_gateway_client::ClientApi;
 use starknet_gateway_types::error::SequencerError;
 use starknet_gateway_types::request::add_transaction::ContractDefinition;
 
-crate::rpc::error::generate_rpc_error_subset!(AddDeclareTransactionError: InvalidContractClass);
+crate::error::generate_rpc_error_subset!(AddDeclareTransactionError: InvalidContractClass);
 
 impl From<SequencerError> for AddDeclareTransactionError {
     fn from(e: SequencerError) -> Self {
@@ -73,8 +73,8 @@ pub async fn add_declare_transaction(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::rpc::v02::types::request::BroadcastedDeclareTransaction;
-    use crate::rpc::v02::types::ContractClass;
+    use crate::v02::types::request::BroadcastedDeclareTransaction;
+    use crate::v02::types::ContractClass;
     use pathfinder_common::{
         starkhash, ContractAddress, Fee, TransactionNonce, TransactionVersion,
     };
